fix(news-hub): address PR #598 R14 — preserve landmark roles, setter pattern, hook utils extraction, require cast comment
